talemeastory/requirements.txt
Django
djangorestframework
psycopg2
gunicorn
python-dotenv
django-cors-headers
boto3
django-storages